Struct chakracore::context::Context

pub struct Context(_);

A sandboxed execution context with its own set of built-in objects and functions.

In a browser or Node.JS environment, the task of executing promises is handled by the runtime. This is not the case with ChakraCore. To run promise chains, execute_tasks must be called at a regular interval. This is done using the ContextGuard.

Methods

impl Context

fn new(runtime: &Runtime) -> Result<Context>

Creates a new context and returns a handle to it.

unsafe fn from_object(object: &Object) -> Result<Context>

Returns an object's associated context.

fn make_current<'a>(&'a self) -> Result<ContextGuard<'a>>

Binds the context to the current scope.

The majority of APIs require an active context.

fn insert_user_data<T>(&self, value: T) -> Option<T> where
T: 'static,

Set user data associated with the context.

Only one value per type.

The internal implementation uses AnyMap.

Returns a previous value if applicable.

The data will live as long as the runtime keeps the context.

fn remove_user_data<T>(&self) -> Option<T> where
T: 'static,

Remove user data associated with the context.

fn get_user_data<T>(&self) -> Option<&T> where
T: 'static,

Get user data associated with the context.

fn get_user_data_mut<T>(&self) -> Option<&mut T> where
T: 'static,

Get mutable user data associated with the context.

unsafe fn get_current<'a>() -> Option<ContextGuard<'a>>

Returns the active context in the current thread.

This is unsafe because there should be little reason to use it in idiomatic code. Usage patterns should utilize ContextGuard instead.

This ContextGuard does not reset the current context upon destruction, in contrast to a normally allocated ContextGuard. This is merely a reference.

fn as_raw(&self) -> JsContextRef

Returns the underlying raw pointer.

impl Context

unsafe fn from_raw(value: JsRef) -> Context

Creates an instance from a raw pointer.

This is used for managing the lifetime of JSRT objects. They are tracked using reference counting; incrementing with from_raw, and decrementing with drop.

This is required to support items stored on the heap, since the JSRT runtime only observes the stack.