Struct chakracore::runtime::Builder

pub struct Builder { /* fields omitted */ }

A builder for the runtime type.

Methods

impl Builder

fn disable_background_work(self) -> Self

Disable the runtime from doing any work on background threads.

fn disable_eval(self) -> Self

Disable eval and function by throwing an exception upon use.

fn disable_jit(self) -> Self

Disable just-in-time compilation.

fn enable_experimental(self) -> Self

Allow experimental JavaScript features.

fn enable_script_interrupt(self) -> Self

Allow script interrupt.

fn dispatch_exceptions(self) -> Self

Dispatch exceptions to any attached JavaScript debuggers.

fn supports_idle_tasks(self) -> Self

Enable idle processing. run_idle_tasks must be called regularly.

fn memory_limit(self, limit: usize) -> Self

Set the runtime's memory limit.

fn collect_callback(self, callback: Box<CollectCallback>) -> Self

Set a callback for when the runtime collects garbage.

fn build(self) -> Result<Runtime>

Creates the runtime object with associated settings.