Struct chakracore::value::Error

pub struct Error(_);

A JavaScript error.

Methods

impl Error

fn new(guard: &ContextGuard, message: &str) -> Self

Creates a new error.

fn range_error(guard: &ContextGuard, message: &str) -> Self

Creates a new range error.

fn reference_error(guard: &ContextGuard, message: &str) -> Self

Creates a new reference error.

fn syntax_error(guard: &ContextGuard, message: &str) -> Self

Creates a new syntax error.

fn type_error(guard: &ContextGuard, message: &str) -> Self

Creates a new type error.

fn uri_error(guard: &ContextGuard, message: &str) -> Self

Creates a new URI error.

fn message(&self, guard: &ContextGuard) -> String

Returns the error's message.

fn is_same(value: &Value) -> bool

Returns true if the value is an Error.

impl Error

unsafe fn from_raw(value: JsRef) -> Error

Creates an instance from a raw pointer.

This is used for managing the lifetime of JSRT objects. They are tracked using reference counting; incrementing with from_raw, and decrementing with drop.

This is required to support items stored on the heap, since the JSRT runtime only observes the stack.