Struct chakracore::value::External

pub struct External(_);

A JavaScript external object.

Methods

impl External

fn new<T>(_guard: &ContextGuard, external: Box<T>) -> Self

Creates a new object with external data.

The object takes ownership of the resource. It is undetermined when, and even if, the destructor is called. It relies on the engine's finalize callback.

As long as the object is referenced on the stack or in any script context, the external data will be kept alive (i.e it is not tied to the handle).

unsafe fn from_ptr<T>(_guard: &ContextGuard, external: *mut T) -> Self

Creates a new object with external data.

This is unsafe because the object does not take ownership of the resource. Therefore the data may become a dangling pointer. The caller is responsible for keeping the reference alive.

unsafe fn value<T>(&self) -> &mut T

Returns the external object's data.

fn is_same(value: &Value) -> bool

Returns true if the value is an External.

impl External

unsafe fn from_raw(value: JsRef) -> External

Creates an instance from a raw pointer.

This is used for managing the lifetime of JSRT objects. They are tracked using reference counting; incrementing with from_raw, and decrementing with drop.

This is required to support items stored on the heap, since the JSRT runtime only observes the stack.

If used in conjunction with a Property or any Value, it is assumed a Context is active.

fn as_raw(&self) -> JsRef

Returns the underlying raw pointer.