Struct chakracore::value::Number

pub struct Number(_);

A JavaScript number.

Methods

impl Number

fn new(_guard: &ContextGuard, number: i32) -> Self

Creates a new number.

fn from_double(_guard: &ContextGuard, number: f64) -> Self

Creates a new number from a double.

fn value_double(&self) -> f64

Converts a JavaScript number to a double.

fn value(&self) -> i32

Converts a JavaScript number to an integral.

fn is_same(value: &Value) -> bool

Returns true if the value is a Number.

impl Number

unsafe fn from_raw(value: JsRef) -> Number

Creates an instance from a raw pointer.

This is used for managing the lifetime of JSRT objects. They are tracked using reference counting; incrementing with from_raw, and decrementing with drop.

This is required to support items stored on the heap, since the JSRT runtime only observes the stack.

If used in conjunction with a Property or any Value, it is assumed a Context is active.

fn as_raw(&self) -> JsRef

Returns the underlying raw pointer.